refactored to fit sidebar into explore
