+++ frontend/components/Layout/LayoutParts/AppFooter.vue
@@ -12,7 +12,7 @@
     </v-row>
   </v-footer>
 </template>
-      
+
 <script lang="ts">
 import { defineComponent } from "@nuxtjs/composition-api";
 
@@ -22,4 +22,3 @@ export default defineComponent({
   },
 });
 </script>
-
