+++ frontend/layouts/admin.vue
@@ -27,9 +27,9 @@
 
 <script lang="ts">
 import { defineComponent, ref, useContext, onMounted } from "@nuxtjs/composition-api";
-import AppHeader from "@/components/Layout/AppHeader.vue";
-import AppSidebar from "@/components/Layout/AppSidebar.vue";
-import TheSnackbar from "~/components/Layout/TheSnackbar.vue";
+import AppHeader from "@/components/Layout/LayoutParts/AppHeader.vue";
+import AppSidebar from "@/components/Layout/LayoutParts/AppSidebar.vue";
+import TheSnackbar from "~/components/Layout/LayoutParts/TheSnackbar.vue";
 import { SidebarLinks } from "~/types/application-types";
 
 export default defineComponent({
@@ -48,28 +48,33 @@ export default defineComponent({
       {
         icon: $globals.icons.cog,
         to: "/admin/site-settings",
-        title: i18n.t("sidebar.site-settings"),
+        title: i18n.tc("sidebar.site-settings"),
+        restricted: true,
       },
 
       // {
       //   icon: $globals.icons.chart,
       //   to: "/admin/analytics",
       //   title: "Analytics",
+      //   restricted: true,
       // },
       {
         icon: $globals.icons.user,
         to: "/admin/manage/users",
-        title: i18n.t("user.users"),
+        title: i18n.tc("user.users"),
+        restricted: true,
       },
       {
         icon: $globals.icons.group,
         to: "/admin/manage/groups",
-        title: i18n.t("group.groups"),
+        title: i18n.tc("group.groups"),
+        restricted: true,
       },
       {
         icon: $globals.icons.database,
         to: "/admin/backups",
-        title: i18n.t("sidebar.backups"),
+        title: i18n.tc("sidebar.backups"),
+        restricted: true,
       },
     ];
 
@@ -77,25 +82,29 @@ export default defineComponent({
       {
         icon: $globals.icons.wrench,
         to: "/admin/maintenance",
-        title: i18n.t("sidebar.maintenance"),
+        title: i18n.tc("sidebar.maintenance"),
+        restricted: true,
       },
       {
         icon: $globals.icons.check,
         to: "/admin/background-tasks",
-        title: i18n.t("sidebar.background-tasks"),
+        title: i18n.tc("sidebar.background-tasks"),
+        restricted: true,
       },
       {
         icon: $globals.icons.slotMachine,
         to: "/admin/parser",
-        title: i18n.t("sidebar.parser"),
+        title: i18n.tc("sidebar.parser"),
+        restricted: true,
       },
     ];
 
     const bottomLinks: SidebarLinks = [
       {
         icon: $globals.icons.heart,
-        title: i18n.t("about.support"),
+        title: i18n.tc("about.support"),
         href: "https://github.com/sponsors/hay-kot",
+        restricted: true,
       },
     ];
