+++ frontend/layouts/basic.vue
@@ -10,12 +10,12 @@
     </v-main>
   </v-app>
 </template>
-  
+
 
 <script lang="ts">
 import { defineComponent } from "@nuxtjs/composition-api";
-import TheSnackbar from "~/components/Layout/TheSnackbar.vue";
-import AppHeader from "@/components/Layout/AppHeader.vue";
+import TheSnackbar from "~/components/Layout/LayoutParts/TheSnackbar.vue";
+import AppHeader from "@/components/Layout/LayoutParts/AppHeader.vue";
 export default defineComponent({
   components: { TheSnackbar, AppHeader },
 });
