+++ frontend/layouts/explore.vue
@@ -1,0 +1,13 @@
+
+<template>
+    <DefaultLayout />
+  </template>
+
+  <script lang="ts">
+  import DefaultLayout from "@/components/Layout/DefaultLayout.vue";
+  import { defineComponent } from "@nuxtjs/composition-api";
+
+  export default defineComponent({
+    components: { DefaultLayout },
+  });
+  </script>
