+++ frontend/types/application-types.ts
@@ -4,8 +4,9 @@ export interface SideBarLink {
   icon: string;
   to?: string;
   href?: string;
-  title: TranslateResult;
+  title: string;
   children?: SideBarLink[];
+  restricted: boolean;
 }
 
 export type SidebarLinks = Array<SideBarLink>;
